Add : forum 생성 api query 추가
